+++ apollo-biz/src/test/java/com/ctrip/framework/apollo/biz/config/BizConfigTest.java
@@ -18,13 +18,13 @@ package com.ctrip.framework.apollo.biz.config;
 
 import com.ctrip.framework.apollo.biz.repository.ServerConfigRepository;
 import com.ctrip.framework.apollo.biz.service.BizDBPropertySource;
+import java.util.Map;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.core.env.ConfigurableEnvironment;
-import org.springframework.core.env.Environment;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import javax.sql.DataSource;
@@ -93,7 +93,7 @@ public class BizConfigTest {
     int someOverrideLimit = 10;
     String overrideValueString = "{'a+b+c+b':10}";
     when(environment.getProperty("apollo.release-history.retention.size.override")).thenReturn(overrideValueString);
-    int  overrideValue = bizConfig.releaseHistoryRetentionSizeOverride().get("a+b+c+b");
+    int overrideValue = bizConfig.releaseHistoryRetentionSizeOverride().get("a+b+c+b");
     assertEquals(someOverrideLimit, overrideValue);
 
     overrideValueString = "{'a+b+c+b':0,'a+b+d+b':2}";
@@ -105,6 +105,48 @@ public class BizConfigTest {
     overrideValueString = "{}";
     when(environment.getProperty("apollo.release-history.retention.size.override")).thenReturn(overrideValueString);
     assertEquals(0, bizConfig.releaseHistoryRetentionSizeOverride().size());
+  }
+
+  @Test
+  public void testAppIdValueLengthLimitOverride() {
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(null);
+    Map<String, Integer> result = bizConfig.appIdValueLengthLimitOverride();
+    assertTrue(result.isEmpty());
+
+    String input = "{}";
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(input);
+    result = bizConfig.appIdValueLengthLimitOverride();
+    assertTrue(result.isEmpty());
+
+    input = "invalid json";
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(input);
+    result = bizConfig.appIdValueLengthLimitOverride();
+    assertTrue(result.isEmpty());
+
+    input = "{'appid1':555}";
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(input);
+    int overrideValue = bizConfig.appIdValueLengthLimitOverride().get("appid1");
+    assertEquals(1, bizConfig.appIdValueLengthLimitOverride().size());
+    assertEquals(555, overrideValue);
+
+    input = "{'appid1':555,'appid2':666}";
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(input);
+    overrideValue = bizConfig.appIdValueLengthLimitOverride().get("appid2");
+    assertEquals(2, bizConfig.appIdValueLengthLimitOverride().size());
+    assertEquals(666, overrideValue);
+
+    input = "{'appid1':555,'appid2':666,'appid3':0,'appid4':-1}";
+    when(environment.getProperty("appid.value.length.limit.override")).thenReturn(input);
+    result = bizConfig.appIdValueLengthLimitOverride();
+
+    assertTrue(result.containsKey("appid1"));
+    assertTrue(result.containsKey("appid2"));
+    assertFalse(result.containsKey("appid3"));
+    assertFalse(result.containsKey("appid4"));
+    assertEquals(2, result.size());
+
+    overrideValue = result.get("appid2");
+    assertEquals(666, overrideValue);
   }
 
   @Test